+++ src/request_response/mod.rs
@@ -67,7 +67,7 @@ mod tests {
     #[should_panic(expected = "not expected")]
     fn request_response_contains_response_request_panics() {
         let target =
-            RequestResponse::<Randoms, RandomsAddRequest>::Response(AddResponse::new(0, Ok(())));
+            RequestResponse::<Randoms, RandomsAddRequest>::Response(AddResponse::new(0, Ok(0)));
 
         target.request();
     }