+++ CHANGELOG.md
@@ -38,4 +38,5 @@ The PoolItem trait provides a cleaner and hopefully more intuitive interface wit
 - Ignore requests to add and existing key; return appropriate error in response
 - Add default implementation of send_and_receive_one to SenderAndReceiver trait
 - Add is_complete function to mocks
+- Add id to the result of AddResponse